LayoutTests/editing/selection/ios/select-content-inside-and-outside-multicolumn-expected.txt

When selecting multi-column content and others, the highlight boxes should not be coalesced.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectionRects.length is 5
PASS selectionRects[0].width is 150
PASS selectionRects[1].width is 150
PASS selectionRects[2].width is 150
PASS selectionRects[3].width is 310
PASS outsideContentLineTop is > lastMultiColumnLineBottom
PASS successfullyParsed is true

TEST COMPLETE
Lorem ipsum dolor sit amet, consectetur adipiscing elit.Morbi mauris ex, finibus quis aliquam ac, interdum facilisis ligula.
The selected content out of the multi-column element.
